[TASK] Add mimetype mpeg/audio to list of allowed mimetypes for solr cell
This pr:

* Add's the mimetype ```mpeg/audio``` as allowed mimetype for solr cell extraction
* Add's a functional test for that and fixes a bug with the extraction query

Fixes: #57
